+++ packages/grafana-ui/src/components/TimePicker/TimePicker.tsx
@@ -1,12 +1,12 @@
 import React, { PureComponent } from 'react';
 import moment from 'moment';
 import { TimeRange, TimeOptions, TimeOption, SelectOptionItem } from '@grafana/ui';
-import { ButtonSelect } from '@grafana/ui/src/components/Select/ButtonSelect';
+import { ButtonSelect } from '../Select/ButtonSelect';
 import { mapTimeOptionToTimeRange, mapTimeRangeToRangeString } from './time';
 import { Props as TimePickerPopoverProps } from './TimePickerPopover';
 import { TimePickerOptionGroup } from './TimePickerOptionGroup';
-import { PopperContent } from '@grafana/ui/src/components/Tooltip/PopperController';
-import { Timezone } from '../../../../../public/app/core/utils/datemath';
+import { PopperContent } from '../Tooltip/PopperController';
+import { Timezone } from '../../utils/datemath';
 
 export interface Props {
   value: TimeRange;
