+++ packages/grafana-ui/src/components/TimePicker/TimePickerPopover.tsx
@@ -1,11 +1,11 @@
 import React, { Component, SyntheticEvent } from 'react';
-import { TimeRange, TimeOptions, TimeOption } from '@grafana/ui';
+import { TimeRange, TimeOptions, TimeOption } from '../../types/time';
 import { Moment } from 'moment';
 
 import { TimePickerCalendar } from './TimePickerCalendar';
 import { TimePickerInput } from './TimePickerInput';
 import { mapTimeOptionToTimeRange } from './time';
-import { Timezone } from '../../../../../public/app/core/utils/datemath';
+import { Timezone } from '../../utils/datemath';
 
 export interface Props {
   value: TimeRange;
