+++ packages/grafana-ui/src/components/TimePicker/time.ts
@@ -1,8 +1,8 @@
 import moment, { Moment } from 'moment';
 import { TimeOption, TimeRange, TIME_FORMAT } from '@grafana/ui';
 
-import * as dateMath from '../../../../../public/app/core/utils/datemath';
-import { describeTimeRange } from '../../../../../public/app/core/utils/rangeutil';
+import * as dateMath from '@grafana/ui/src/utils/datemath';
+import { describeTimeRange } from '@grafana/ui/src/utils/rangeutil';
 
 export const mapTimeOptionToTimeRange = (
   timeOption: TimeOption,
