+++ packages/grafana-ui/src/utils/rangeutil.ts
