+++ public/app/features/dashboard/components/TimePicker/TimePickerCtrl.ts
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import angular from 'angular';
 import moment from 'moment';
 
-import * as rangeUtil from 'app/core/utils/rangeutil';
+import * as rangeUtil from '@grafana/ui/src/utils/rangeutil';
 
 export class TimePickerCtrl {
   static tooltipFormat = 'MMM D, YYYY HH:mm:ss';
