+++ public/app/features/dashboard/components/TimePicker/validation.ts
@@ -1,5 +1,5 @@
 import moment from 'moment';
-import * as dateMath from 'app/core/utils/datemath';
+import * as dateMath from '@grafana/ui/src/utils/datemath';
 
 export function inputDateDirective() {
   return {
