+++ public/app/features/dashboard/services/DashboardLoaderSrv.ts
@@ -3,7 +3,7 @@ import moment from 'moment';
 import _ from 'lodash';
 import $ from 'jquery';
 import kbn from 'app/core/utils/kbn';
-import * as dateMath from 'app/core/utils/datemath';
+import * as dateMath from '@grafana/ui/src/utils/datemath';
 import impressionSrv from 'app/core/services/impression_srv';
 
 export class DashboardLoaderSrv {
