+++ public/app/plugins/datasource/cloudwatch/specs/datasource.test.ts
@@ -1,6 +1,6 @@
 import '../datasource';
 import CloudWatchDatasource from '../datasource';
-import * as dateMath from 'app/core/utils/datemath';
+import * as dateMath from '@grafana/ui/src/utils/datemath';
 import { TemplateSrv } from 'app/features/templating/template_srv';
 import { CustomVariable } from 'app/features/templating/all';
 import _ from 'lodash';
