+++ public/test/specs/helpers.ts
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import config from 'app/core/config';
-import * as dateMath from 'app/core/utils/datemath';
+import * as dateMath from '@grafana/ui/src/utils/datemath';
 import { angularMocks, sinon } from '../lib/common';
 import { PanelModel } from 'app/features/dashboard/state/PanelModel';
 import { PanelPluginMeta, RawTimeRange } from '@grafana/ui';
